requirements.txt
numpy
pandas
matplotlib
praw
s3fs
openpyxl
apache-airflow-providers-amazon